chore: make clippy shut up
