+++ zkdoc_sdk/src/circuits/file_hash_partial_v3.rs
@@ -1,15 +1,16 @@
 use crate::gadgets::{
-    file_hash_row_selector::{self, FileHashRowSelectorChip, FileHashRowSelectorConfig},
+    file_hash_row_selector::{FileHashRowSelectorChip, FileHashRowSelectorConfig},
     file_selector_accumulator::{FileSelectorAccumulatorChip, FileSelectorAccumulatorConfig},
     poseidon::{PoseidonChip, PoseidonConfig},
 };
-use halo2_gadgets::{poseidon::primitives::P128Pow5T3, utilities::Var};
+use halo2_gadgets::poseidon::primitives::P128Pow5T3;
 use halo2_proofs::{
-    circuit::{floor_planner::V1, Layouter, Value},
+    circuit::{floor_planner::V1, Value},
     pasta::Fp,
-    plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Instance},
+    plonk::{Advice, Circuit, Column, ConstraintSystem, Instance},
 };
 
+#[allow(unused)]
 #[derive(Clone)]
 pub struct FileHashPartialConfig {
     row_title_advice: Column<Advice>,
@@ -21,6 +22,7 @@ pub struct FileHashPartialConfig {
     row_selector_accumulator: FileSelectorAccumulatorConfig,
 }
 
+#[allow(unused)]
 pub struct FileHashPartialChip {
     config: FileHashPartialConfig,
 }
@@ -39,8 +41,8 @@ impl FileHashPartialChip {
         meta.enable_equality(instance);
 
         let state = (0..3).map(|_| meta.advice_column()).collect::<Vec<_>>();
-        for i in 0..3 {
-            meta.enable_equality(state[i]);
+        for i in state.iter().take(3) {
+            meta.enable_equality(*i);
         }
 
         let poseidon_config = PoseidonChip::<P128Pow5T3, 3, 2, 2>::configure(meta, state);
@@ -116,10 +118,10 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
         let row_selector_cs = FileHashRowSelectorChip::<Fp>::construct(config.row_selector_config);
 
         let mut selected_rows = Vec::new();
-        for i in 0..(file_hashes.len()) {
+        for (i, hash) in file_hashes.iter().enumerate() {
             let (file_hash_cell, _, file_res_cell) = row_selector_cs.assign(
                 layouter.namespace(|| "row selectors"),
-                file_hashes[i].value().copied(),
+                hash.value().copied(),
                 self.row_selectors[i],
                 i,
             )?;
@@ -154,11 +156,11 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
             },
         )?;
 
-        for i in 2..L {
+        for (i, selected) in selected_rows.iter().enumerate().take(L).skip(2) {
             let (b_cell, res_cell) = row_selector_accumulator_cs.assign(
                 layouter.namespace(|| "row selector rest of the rows"),
                 &row_selector_accumulator,
-                selected_rows[i].value().copied(),
+                selected.value().copied(),
                 i - 1,
             )?;
 
@@ -184,8 +186,8 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
             &starting_poseidon_hash_message,
         )?;
 
-        for i in 2..L {
-            let message_cells = [accumulated_hash.clone(), file_hashes[i].clone()];
+        for hash in file_hashes.iter().take(L).skip(2) {
+            let message_cells = [accumulated_hash.clone(), hash.clone()];
             accumulated_hash =
                 poseidon_cs.hash(layouter.namespace(|| "poseidon chip"), &message_cells)?;
         }
