+++ zkdoc_sdk/src/circuits/file_hasher.rs
@@ -1,13 +1,13 @@
 use crate::gadgets::{
-    file_hash_row_selector::{self, FileHashRowSelectorChip, FileHashRowSelectorConfig},
+    file_hash_row_selector::{FileHashRowSelectorChip, FileHashRowSelectorConfig},
     file_selector_accumulator::{FileSelectorAccumulatorChip, FileSelectorAccumulatorConfig},
     poseidon::{PoseidonChip, PoseidonConfig},
 };
-use halo2_gadgets::{poseidon::primitives::P128Pow5T3, utilities::Var};
+use halo2_gadgets::poseidon::primitives::P128Pow5T3;
 use halo2_proofs::{
-    circuit::{floor_planner::V1, Layouter, Value},
+    circuit::{floor_planner::V1, Value},
     pasta::Fp,
-    plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Instance},
+    plonk::{Circuit, Column, ConstraintSystem, Instance},
 };
 
 #[derive(Clone)]
@@ -18,6 +18,7 @@ pub struct FileHashPartialConfig {
     row_selector_accumulator: FileSelectorAccumulatorConfig,
 }
 
+#[allow(unused)]
 pub struct FileHashPartialChip {
     config: FileHashPartialConfig,
 }
@@ -33,8 +34,8 @@ impl FileHashPartialChip {
         meta.enable_equality(instance);
 
         let state = (0..3).map(|_| meta.advice_column()).collect::<Vec<_>>();
-        for i in 0..3 {
-            meta.enable_equality(state[i]);
+        for state in state.iter().take(3) {
+            meta.enable_equality(*state);
         }
 
         let poseidon_config = PoseidonChip::<P128Pow5T3, 3, 2, 2>::configure(meta, state);
@@ -169,10 +170,10 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
         let row_selector_cs = FileHashRowSelectorChip::<Fp>::construct(config.row_selector_config);
 
         let mut selected_rows = Vec::new();
-        for i in 0..(file_hashes.len()) {
+        for (i, hash) in file_hashes.iter().enumerate() {
             let (file_hash_cell, _, file_res_cell) = row_selector_cs.assign(
                 layouter.namespace(|| "row selectors"),
-                file_hashes[i].value().copied(),
+                hash.value().copied(),
                 self.row_selectors[i],
                 i,
             )?;
@@ -207,11 +208,11 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
             },
         )?;
 
-        for i in 2..L {
+        for (i, selected) in selected_rows.iter().enumerate().take(L).skip(2) {
             let (b_cell, res_cell) = row_selector_accumulator_cs.assign(
                 layouter.namespace(|| "row selector rest of the rows"),
                 &row_selector_accumulator,
-                selected_rows[i].value().copied(),
+                selected.value().copied(),
                 i - 1,
             )?;
 
@@ -237,8 +238,8 @@ impl<const L: usize> Circuit<Fp> for FileHashPartialCircuit<L> {
             &starting_poseidon_hash_message,
         )?;
 
-        for i in 2..L {
-            let message_cells = [accumulated_hash.clone(), file_hashes[i].clone()];
+        for hash in file_hashes.into_iter().take(L).skip(2) {
+            let message_cells = [accumulated_hash.clone(), hash];
             accumulated_hash =
                 poseidon_cs.hash(layouter.namespace(|| "poseidon chip"), &message_cells)?;
         }
@@ -262,8 +263,9 @@ mod tests {
     };
     use rand_core::OsRng;
 
+    use crate::utils::conversion::convert_hash_u32_to_u64;
+
     use super::FileHashPartialCircuit;
-    use crate::utils::poseidon::convert_hash_u32_to_u64;
     use std::time::Instant;
 
     #[test]
@@ -341,6 +343,7 @@ mod tests {
         prover.assert_satisfied();
     }
 
+    // testing only 2 rows
     #[test]
     fn test_real_prover() {
         let row_selector = [Fp::from(1), Fp::from(0)];
@@ -359,12 +362,6 @@ mod tests {
 
         let row_title = [row_title, row_title];
         let row_content = [row_content, row_content];
-
-        let circuit = FileHashPartialCircuit::<2> {
-            row_title: row_title.map(|x| x.map(|y| Value::known(y))),
-            row_content: row_content.map(|x| x.map(|y| Value::known(y))),
-            row_selectors: row_selector.map(|x| Value::known(x)),
-        };
 
         let mut row_hash = Vec::new();
         let mut row_accumulator = Fp::zero();
@@ -409,8 +406,8 @@ mod tests {
         let mut accumulator_hash = poseidon::Hash::<_, P128Pow5T3, ConstantLength<2>, 3, 2>::init()
             .hash([row_hash[0], row_hash[1]]);
 
-        for i in 2..row_content.len() {
-            let message = [accumulator_hash, row_hash[i]];
+        for i in row_hash.into_iter().skip(2) {
+            let message = [accumulator_hash, i];
             let output =
                 poseidon::Hash::<_, P128Pow5T3, ConstantLength<2>, 3, 2>::init().hash(message);
             accumulator_hash = output;
@@ -494,12 +491,6 @@ mod tests {
         xx.append(&mut yy);
         let row_selector: [Fp; ROW_NUMBER] = xx.try_into().unwrap();
 
-        let circuit = FileHashPartialCircuit::<ROW_NUMBER> {
-            row_title: row_title.map(|x| x.map(|y| Value::known(y))),
-            row_content: row_content.map(|x| x.map(|y| Value::known(y))),
-            row_selectors: row_selector.map(|x| Value::known(x)),
-        };
-
         let mut row_hash = Vec::new();
         let mut row_accumulator = Fp::zero();
         for ((&title, &content), &row_selector) in row_title
@@ -543,8 +534,8 @@ mod tests {
         let mut accumulator_hash = poseidon::Hash::<_, P128Pow5T3, ConstantLength<2>, 3, 2>::init()
             .hash([row_hash[0], row_hash[1]]);
 
-        for i in 2..row_content.len() {
-            let message = [accumulator_hash, row_hash[i]];
+        for i in row_hash.into_iter().skip(2) {
+            let message = [accumulator_hash, i];
             let output =
                 poseidon::Hash::<_, P128Pow5T3, ConstantLength<2>, 3, 2>::init().hash(message);
             accumulator_hash = output;
