+++ zkdoc_sdk/src/circuits/poseidon.rs
@@ -12,11 +12,13 @@ struct PoseidonConfig {
     poseidon_config: Config<3, 2, 2>,
 }
 
+#[allow(unused)]
 struct PoseidonChip {
     config: PoseidonConfig,
 }
 
 impl PoseidonChip {
+    #[allow(unused)]
     pub fn construct(config: PoseidonConfig) -> Self {
         Self { config }
     }
@@ -26,8 +28,8 @@ impl PoseidonChip {
         meta.enable_equality(instance);
 
         let state = (0..3).map(|_| meta.advice_column()).collect::<Vec<_>>();
-        for i in 0..3 {
-            meta.enable_equality(state[i]);
+        for i in state.iter().take(3) {
+            meta.enable_equality(*i);
         }
 
         let poseidon_config = Chip::<P128Pow5T3, 3, 2, 2>::configure(meta, state);
@@ -79,52 +81,12 @@ impl Circuit<Fp> for PoseidonCircuit {
 #[cfg(test)]
 mod tests {
     use ff::PrimeFieldBits;
-    use halo2_gadgets::poseidon::{
-        primitives::{self as poseidon, ConstantLength, P128Pow5T3, Spec},
-        Hash,
-    };
+    use halo2_gadgets::poseidon::primitives::{self as poseidon, ConstantLength, P128Pow5T3};
     use halo2_proofs::{arithmetic::FieldExt, circuit::Value, dev::MockProver, pasta::Fp};
 
+    use crate::utils::conversion::{convert_hash_u32_to_u64, convert_hash_u64_to_u32};
+
     use super::PoseidonCircuit;
-
-    fn convert_hash_u64_to_u32(hash_u64: [u64; 4]) -> [u32; 8] {
-        let mut res = Vec::new();
-        for num in hash_u64 {
-            let res_u32 = convert_u64_to_u32_BE(num);
-            res.push(res_u32[0]);
-            res.push(res_u32[1]);
-        }
-
-        res.try_into().unwrap()
-    }
-
-    fn convert_u64_to_u32_BE(input: u64) -> [u32; 2] {
-        let lower = input as u32;
-        let upper = (input >> 32) as u32;
-
-        [upper, lower]
-    }
-
-    fn convert_hash_u32_to_u64(hash_u32: [u32; 8]) -> [u64; 4] {
-        let mut res = Vec::new();
-        for i in 0..4 {
-            let starting_index = i * 2;
-            let arr = [hash_u32[starting_index], hash_u32[starting_index + 1]];
-            let res_u64 = convert_u32_to_u64_BE(arr);
-            res.push(res_u64);
-        }
-
-        res.reverse();
-        res.try_into().unwrap()
-    }
-
-    fn convert_u32_to_u64_BE(u32_array: [u32; 2]) -> u64 {
-        u32_array[0] as u64 * u64::pow(2, 32) + u32_array[1] as u64
-    }
-
-    fn convert_u32_to_u64_LE(u32_array: [u32; 2]) -> u64 {
-        u32_array[1] as u64 * u64::pow(2, 32) + u32_array[0] as u64
-    }
 
     #[test]
     fn test_convert() {
@@ -182,10 +144,6 @@ mod tests {
             "output bits raw slice u32: {:?}",
             convert_hash_u64_to_u32(output.to_le_bits().as_raw_slice().try_into().unwrap())
         );
-        let output_le_bits2 = output.to_le_bits();
-        let output_le_bits = output.to_le_bits().as_raw_slice();
-
-        let test_fp = Fp::from_raw([1, 3, 4, 5]);
 
         let pub_instance = vec![output];
