+++ zkdoc_sdk/src/gadgets/file_hash_row_selector.rs
@@ -61,6 +61,7 @@ impl<F: FieldExt> FileHashRowSelectorChip<F> {
         }
     }
 
+    #[allow(clippy::type_complexity)]
     pub fn assign(
         &self,
         mut layouter: impl Layouter<F>,
