+++ zkdoc_sdk/src/gadgets/poseidon_v2.rs
@@ -43,8 +43,8 @@ impl<S: Spec<Fp, WIDTH, RATE>, const WIDTH: usize, const RATE: usize, const L: u
         let rc_a = (0..WIDTH).map(|_| meta.fixed_column()).collect::<Vec<_>>();
         let rc_b = (0..WIDTH).map(|_| meta.fixed_column()).collect::<Vec<_>>();
         let instance = meta.instance_column();
-        for i in 0..WIDTH {
-            meta.enable_equality(state[i]);
+        for i in state.iter().take(WIDTH) {
+            meta.enable_equality(*i);
         }
         meta.enable_equality(instance);
         meta.enable_constant(rc_b[0]);
@@ -52,15 +52,15 @@ impl<S: Spec<Fp, WIDTH, RATE>, const WIDTH: usize, const RATE: usize, const L: u
         let pow5_config = Pow5Chip::configure::<S>(
             meta,
             state.clone().try_into().unwrap(),
-            partial_sbox.try_into().unwrap(),
+            partial_sbox,
             rc_a.try_into().unwrap(),
             rc_b.try_into().unwrap(),
         );
 
         PoseidonConfig {
-            inputs: state.clone().try_into().unwrap(),
+            inputs: state,
             instance,
-            pow5_config: pow5_config,
+            pow5_config,
         }
     }
 
@@ -112,7 +112,7 @@ impl<S: Spec<Fp, WIDTH, RATE>, const WIDTH: usize, const RATE: usize, const L: u
                     .enumerate()
                     .map(|(i, word)| {
                         word.copy_advice(
-                            || format!("word {}", i),
+                            || format!("word {i}"),
                             &mut region,
                             self.config.inputs[i],
                             0,
